refactor: cart api 提取可重用功能，取得課程資訊與計算課程總價
cartOperations 取得課程資料與課程總價
